==> Python_data_analysis_project/car_insurance_pricing_analytics/notebooks/01_policy_data_generation_cleaning.ipynb <==
import pandas as pd 
import numpy as np 


np.random.seed(42)

start_date = "2023-01-01"
end_date = "2024-12-31"
dates= pd.date_range(start = start_date,end = end_date,freq = "D")

num_customers = 3000

vehicle_types = ["Hatchback", "Sedan", "SUV", "Luxury"]
policy_types = ["Third Party", "Comprehensive"]

regions = ["North", "South", "East", "West"]




data = []
for i in range(num_customers):
    policy_id =f"POL{i+10001}"
    
    start_day = np.random.choice(dates)
    policy_term_days = 365

    vehicle_type = np.random.choice(vehicle_types, p=[0.35, 0.30, 0.25, 0.10])
    policy_type = np.random.choice(policy_types, p=[0.4, 0.6])
    region = np.random.choice(regions)

    vehicle_age = np.random.randint(0, 15)
    driver_age = np.random.randint(21, 65)

    base_premium = {
        "Hatchback": 8000,
        "Sedan": 10000,
        "SUV": 13000,
        "Luxury": 18000
    }[vehicle_type]

    # Risk adjustments
    premium = (
        base_premium
        + vehicle_age * 300
        + (driver_age < 25) * 2000
        + (policy_type == "Comprehensive") * 2500
    )

    discount_pct = np.random.choice(
        [0, 5, 10, 15],
        p=[0.45, 0.25, 0.20, 0.10]
    )

    final_premium = premium * (1 - discount_pct / 100)

    expected_claim_cost = final_premium * np.random.uniform(0.4, 0.8)


    data.append([
        policy_id,
        start_day,
        vehicle_type,
        policy_type,
        region,
        vehicle_age,
        driver_age,
        premium,
        discount_pct,
        final_premium,
        expected_claim_cost
    ])




df = pd.DataFrame(
    data ,
    columns = [
        "policy_id",
        "policy_start_date",
        "vehicle_type",
        "policy_type",
        "region",
        "vehicle_age",
        "driver_age",
        "base_premium",
        "discount_pct",
        "final_premium",
        "expected_claim_cost"
    ]
)

df.head()

df.info()

df.isna().sum()

# Remove unrealistic ages
df = df[(df["driver_age"] >= 18) & (df["driver_age"] <= 75)]

# Remove negative premiums (safety check)
df = df[df["final_premium"] > 0]
df.head()

df["policy_year"] =df["policy_start_date"].dt.year
df.head()

df["underwriting_margin"] = (
    df["final_premium"] - df["expected_claim_cost"]
)
df.head()

# Save RAW (reference)
df.to_csv(
    "C:/Users/Abhi/Desktop/Python_Data_Analytics_Projects/car_insurance_pricing_analytics/data/raw/car_insurance_policy_data.csv",
    index=False
)

df.to_csv(
        "C:/Users/Abhi/Desktop/Python_Data_Analytics_Projects/car_insurance_pricing_analytics/data/processed/car_insurance_policy_cleaned.csv",
    index=False
)

df.describe()
