==> Python_data_analysis_project/car_insurance_pricing_analytics/notebooks/02_premium_discount_feature_engineering.ipynb <==
import pandas as pd
import numpy as np

df = pd.read_csv(
     "C:/Users/Abhi/Desktop/Python_Data_Analytics_Projects/car_insurance_pricing_analytics/data/processed/car_insurance_policy_cleaned.csv",
    parse_dates=["policy_start_date"]
)

df.head()

df["premium_per_vehicle_age"] =(
    df["final_premium"] /(df["vehicle_age"] + 1)
)
df.head()

df["discount_applied"]  = np.where(
    df["discount_pct"] > 0, 1, 0
)   

df["discount_bucket"] = pd.cut(
    df["discount_pct"],
    bins=[-1, 0, 5, 10, 15],
    labels=["No Discount", "Low", "Medium", "High"]
)
df.head()

df["avg_vehicle_premium"] = (
    df.groupby("vehicle_type")["final_premium"]
      .transform("mean")
)
df.head()

df["premium_deviation_pct"] = (
    (df["final_premium"] - df["avg_vehicle_premium"])
    / df["avg_vehicle_premium"]
) * 100
df.head()


df["risk_adjusted_discount"] = (
    df["discount_pct"] / (df["vehicle_age"] + 1)
)
df.head()

df["pricing_status"] = np.where(
    df["underwriting_margin"] < 0,
    "Underpriced",
    "Profitable"
)
df.head()

df["effective_price_index"] = (
    df["final_premium"] / df["avg_vehicle_premium"]
)
df.head()

df.to_csv(
    "C:/Users/Abhi/Desktop/Python_Data_Analytics_Projects/car_insurance_pricing_analytics/data/processed/car_insurance_pricing_features.csv",
    index=False
)

df[
    [
        "vehicle_type",
        "final_premium",
        "discount_pct",
        "discount_bucket",
        "premium_deviation_pct",
        "pricing_status"
    ]
].head()


df["discount_bucket"].value_counts()
